Moved non-test RSpec files to /test/support (RSpec convention)
