+++ test/boot_requirements_checker_s390_test.rb
@@ -21,7 +21,7 @@
 # find current contact information at www.suse.com.
 
 require_relative "spec_helper"
-require_relative "boot_requirements_context"
+require_relative "support/boot_requirements_context"
 require "storage/proposal"
 require "storage/boot_requirements_checker"
 require "storage/refinements/size_casts"
