+++ test/spec_helper.rb
@@ -22,8 +22,8 @@ DATA_PATH = File.expand_path("../data", __FILE__)
 ENV["Y2DIR"] = SRC_PATH
 
 require "yast"
-require_relative "storage_matchers"
-require_relative "storage_helpers"
+require_relative "support/storage_matchers"
+require_relative "support/storage_helpers"
 
 RSpec.configure do |c|
   c.include Yast::RSpec::StorageMatchers
