+++ test/support/proposed_boot_partition_examples.rb
